Исправлены ошибки CheckStyle в Task.java
