+++ src/Main.java
@@ -97,8 +97,3 @@ public class Main {
         System.out.println("-".repeat(100));
     }
 }
-
-
-
-
-
