+++ src/managers/DoublyLinkedList.java
@@ -4,19 +4,3 @@ public class DoublyLinkedList<T> {
     public Node<T> head;
     public Node<T> tail;
 }
-
-class Node <T> {
-    public T data;
-    public Node<T> next;
-    public Node<T> prev;
-
-    public Node(T data) {
-        this.data = data;
-        this.next = null;
-        this.prev = null;
-    }
-}
-
-
-
-
