+++ src/tasks/Epic.java
@@ -30,12 +30,3 @@ public class Epic extends Task {
         return super.toString();
     }
 }
-
-
-
-
-
-
-
-
-
